fix review UI Delete Chapter
